Add random dist matrix generator
Generates two types of random dist matricies, one is noisy where all the
dist values are random, and the other is clustered points where the user
can specify number of clusters and dimension of the raw data. In
addition, augmented CLI so full functionality of the program is
available without code changes.

- Add MDS graph clique algorithm to remove edges based on a factor to
the original distance

- Add random dist matrix generation

- Add random point generation

- Add functionality to reload previously generated data

- Output data to the out/ directory

- Augment MDS to handle higher dimensional data

- Augment CLI, avoiding hard-code changes