Add stress measure to relative MDS
Stress measure takes into account the position of the basis vectors and
the result of the relative MDS (on the remaining points), and calculates
the final stress for that particular dataset. This is to allow
comparative measurement against MDS result. Also made slight
modification to enable retrieval of stress measure from MDS.